tests: load plugins when for command tester
This change cleans up loading of plugins when testing commands.
